clean up energy distribution code (not working)
